clearledger/server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pyjwt==2.10.1
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openpyxl==3.1.5
python-multipart==0.0.18
pydantic-settings==2.7.1
email-validator>=2.0.0
aiofiles==24.1.0
